fix: 서비스 로직 수정
